Finalize the attachment move to be in `create_volume()`
The move of the remote attachment would fit better after the volumes are
actually switched in `initialize_connection()`, however it needs to be
looked up there as the ID is unknown at that point. Looking up the
attachment by its attributes seems error-prone.
The exact attachment is known in `create_volume()`, so we move it to the
other volume there, before the actual volume switch.

After the volume is created it returns as `available`, therefore we set
it to `in-use`, since we know that it has the remote attachment.

Signed-off-by: Jan Horstmann <horstmann@osism.tech>